# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: DS Education Platform API
  version: 0.1.0
  description: |
    Headless learner API for the DS Education Platform (CME).

    Every request carries two credentials:
    - a **bearer token**, validated server-side against the project's own
      Keycloak realm (signature, issuer, audience, expiry — ADR-0003), and
    - the **X-DS-Project header**, identifying which host surface is calling
      (ADR-0007). The project binding decides the Keycloak realm and pins the
      tenant; a token from another realm is rejected regardless of validity.

    All compliance decisions are server-side. The client renders; it is never
    the source of truth for anything affecting a CME point.
servers:
  - url: /
    description: Relative to wherever the API is deployed.
security:
  - bearerAuth: []
...
paths:
...
  /courses/{slug}/certificate:
    get:
      tags: [certificate]
      operationId: getCertificate
      summary: The certificate's fields
      description: |
        The assembled Teilnahmebescheinigung data, for rendering a preview
        before download. Available only once the course is complete.
      parameters:
        - $ref: "#/components/parameters/ProjectHeader"
        - $ref: "#/components/parameters/CourseSlug"
      responses:
        "200":
          description: The certificate data.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Certificate" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404":
          description: Not enrolled, or the course is not complete.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
  /courses/{slug}/certificate/pdf:
    get:
      tags: [certificate]
      operationId: downloadCertificate
      summary: Download the Teilnahmebescheinigung as a PDF
      description: |
        Renders the ÄKWL Muster: participant, course, both VNR barcodes
        (Code 39 and Datamatrix), the participation date, the creditability
        sentence, and the stamp and signature of the course's Wissenschaftliche
        Leitung.

        Refuses with 409 if the course has no stamp or signature configured —
        the Bescheid makes a certificate without them invalid, and issuing an
        invalid-looking document to a physician who has earned their points is
        worse than telling the admin their course is incomplete.
      parameters:
        - $ref: "#/components/parameters/ProjectHeader"
        - $ref: "#/components/parameters/CourseSlug"
      responses:
        "200":
          description: The certificate.
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: |
            Not complete, or the course is missing its signing assets.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Keycloak access token. Verified against the realm bound to the
        X-DS-Project header — issuer and audience come from that binding,
        never from the token (ADR-0003).
  parameters:
    CourseSlug:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
    ProjectHeader:
      name: X-DS-Project
      in: header
      required: true
      description: |
        Project slug identifying the calling host surface (ADR-0007). Pins the
        tenant, and on the learner plane also resolves the Keycloak realm to
        validate the bearer token against.

        **How a bad slug is answered depends on which plane asked**, because the
        two callers know different things already (P22-01):

        - *Learner plane* (bearer token): an unknown **or unbound** slug is a
          generic `401`, never a `404` — whether a project exists is not a fact
          an anonymous caller should be able to enumerate, and a project with no
          Keycloak binding cannot authenticate anybody in any case.
        - *Staff plane* (session cookie, ADR-0012): an unknown slug is a `404`
          carrying `detail`. The caller is already authenticated and the
          platform knows who they are, so naming what was not found is both
          honest and safe. A staff session needs no identity provider at all, so
          a project **without** a Keycloak binding resolves normally here —
          answering 401 for that locked operators out of every tenant-scoped
          console screen on a project the console itself had just created.
        - Either plane, **header absent**: `422` with `detail`. The header is
          required; omitting it is a malformed request, not a failed
          authentication, and answering 401 makes a console send the operator
          back to a login form they never left.

        A caller who is authenticated but holds no grant reaching the resolved
        customer gets `403` on both planes.
      schema: { type: string }
  responses:
    Unauthenticated:
      description: |
        Missing/invalid token, missing project header, or unknown project.
        Always the same generic document — the reason is logged server-side
        only.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
    NotFound:
      description: |
        Not found — or not visible in this tenant; the two are
        indistinguishable by design.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807. No stack traces, no internal identifiers.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlationId:
          type: string
          format: uuid
          description: |
            The id under which the API logged this exact failure (P122-01).

            `problem-details.filter.ts` has put this on every error response
            since observability landed, and it was documented in no contract and
            read by no client — so the one string that finds the failing request
            in the server log reached the payload and stopped there. Somebody
            reporting a failure could not hand over the thing that would locate
            it, because nothing ever showed it to them.

            **Safe to display.** A random UUID minted per request. It identifies
            a log line, never a person.
...
    Certificate:
      type: object
      description: |
        Exactly the fields the Bescheid requires a Teilnahmebescheinigung to
        carry, plus the pre-templated creditability sentence so no client ever
        hardcodes a points figure.
      required:
        - vnr
        - courseTitle
        - completedAt
        - eventLocation
        - organizer
        - cmePoints
        - cmeCategory
        - accreditationBody
        - participantName
        - scientificLeadName
        - creditSentence
      properties:
        vnr: { type: string }
        courseTitle: { type: string }
        completedAt:
          type: string
          format: date-time
          description: |
            Both the Veranstaltungs-datum and -uhrzeit. For an on-demand course
            this is the moment the learner completed it — the same instant that
            starts the EIV reporting window, so certificate and Punktemeldung
            agree by construction.
        eventLocation: { type: string }
        organizer: { type: string }
        cmePoints: { type: integer }
        cmeCategory: { type: string }
        accreditationBody: { type: string }
        participantName: { type: string }
        participantAddress:
          type: string
          description: |
            On the Muster but not in the Bescheid's minimum list, so the
            platform does not collect it and the line renders blank.
        scientificLeadName: { type: string }
        creditSentence: { type: string }
...